# api/resources/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "summary": "Cadastrar usuário (US-03)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão (apenas administrador)"
          },
          "409": {
            "description": "E-mail já cadastrado"
          }
        },
        "tags": [
          "Usuários"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Listar usuários",
        "parameters": [
          {
            "name": "perfil",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ADMIN",
                "AVALIADOR",
                "AVALIADO"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Usuários"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Detalhar usuário",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "Usuários"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/questions": {
      "post": {
        "operationId": "QuestionsController_create",
        "summary": "Cadastrar pergunta aberta ou fechada (US-04/US-05)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos (peso, alternativas, gabarito)"
          },
          "403": {
            "description": "Sem permissão (apenas avaliador)"
          }
        },
        "tags": [
          "Perguntas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "QuestionsController_findAll",
        "summary": "Listar e filtrar perguntas (US-06)",
        "parameters": [
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ABERTA",
                "FECHADA"
              ],
              "type": "string"
            }
          },
          {
            "name": "busca",
            "required": false,
            "in": "query",
            "description": "Palavra no enunciado",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Perguntas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/questions/{id}": {
      "get": {
        "operationId": "QuestionsController_findOne",
        "summary": "Detalhar pergunta",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Pergunta não encontrada"
          }
        },
        "tags": [
          "Perguntas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "QuestionsController_update",
        "summary": "Editar pergunta (US-07)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Pergunta em processo ativo não pode ser editada"
          },
          "404": {
            "description": "Pergunta não encontrada"
          }
        },
        "tags": [
          "Perguntas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "QuestionsController_remover",
        "summary": "Excluir pergunta não vinculada (US-07)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pergunta excluída"
          },
          "400": {
            "description": "Pergunta vinculada a uma avaliação"
          },
          "404": {
            "description": "Pergunta não encontrada"
          }
        },
        "tags": [
          "Perguntas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "API — Software de Processo Avaliativo",
    "description": "API REST para condução de processos avaliativos (avaliador e avaliado). Autenticação via JWT no cabeçalho Authorization: Bearer <token>.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Maria Avaliadora"
          },
          "email": {
            "type": "string",
            "example": "maria@avaliacao.local"
          },
          "senha": {
            "type": "string",
            "example": "Senha@123",
            "minLength": 6
          },
          "perfil": {
            "type": "string",
            "enum": [
              "ADMIN",
              "AVALIADOR",
              "AVALIADO"
            ],
            "example": "AVALIADO"
          },
          "telefone": {
            "type": "string",
            "example": "+5511999998888",
            "description": "Celular do avaliado (opcional)"
          }
        },
        "required": [
          "nome",
          "email",
          "senha",
          "perfil"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "ADMIN",
              "AVALIADOR",
              "AVALIADO"
            ]
          },
          "telefone": {
            "type": "object",
            "nullable": true
          },
          "criadoEm": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "perfil",
          "criadoEm"
        ]
      },
      "CreateAlternativaDto": {
        "type": "object",
        "properties": {
          "texto": {
            "type": "string",
            "example": "Verdadeiro"
          },
          "correta": {
            "type": "boolean",
            "example": false,
            "description": "Indica se é a alternativa correta (gabarito)"
          }
        },
        "required": [
          "texto",
          "correta"
        ]
      },
      "CreateQuestionDto": {
        "type": "object",
        "properties": {
          "enunciado": {
            "type": "string",
            "example": "Explique o conceito de testes exploratórios."
          },
          "tipo": {
            "type": "string",
            "enum": [
              "ABERTA",
              "FECHADA"
            ],
            "example": "ABERTA"
          },
          "peso": {
            "type": "number",
            "example": 10,
            "description": "Peso em pontos (maior que zero)"
          },
          "instrucoes": {
            "type": "string",
            "example": "Responda em até 10 linhas."
          },
          "alternativas": {
            "description": "Obrigatório e com no mínimo 2 itens para perguntas fechadas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateAlternativaDto"
            }
          }
        },
        "required": [
          "enunciado",
          "tipo",
          "peso"
        ]
      },
      "AlternativaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "texto": {
            "type": "string"
          },
          "correta": {
            "type": "boolean",
            "description": "Indica se é a alternativa correta. Omitido para o avaliado durante a prova."
          }
        },
        "required": [
          "id",
          "texto",
          "correta"
        ]
      },
      "QuestionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "enunciado": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "ABERTA",
              "FECHADA"
            ]
          },
          "peso": {
            "type": "number"
          },
          "instrucoes": {
            "type": "object",
            "nullable": true
          },
          "ativa": {
            "type": "boolean"
          },
          "criadoEm": {
            "format": "date-time",
            "type": "string"
          },
          "alternativas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlternativaResponseDto"
            }
          }
        },
        "required": [
          "id",
          "enunciado",
          "tipo",
          "peso",
          "ativa",
          "criadoEm"
        ]
      },
      "UpdateQuestionDto": {
        "type": "object",
        "properties": {
          "enunciado": {
            "type": "string",
            "example": "Explique o conceito de testes exploratórios."
          },
          "tipo": {
            "type": "string",
            "enum": [
              "ABERTA",
              "FECHADA"
            ],
            "example": "ABERTA"
          },
          "peso": {
            "type": "number",
            "example": 10,
            "description": "Peso em pontos (maior que zero)"
          },
          "instrucoes": {
            "type": "string",
            "example": "Responda em até 10 linhas."
          },
          "alternativas": {
            "description": "Obrigatório e com no mínimo 2 itens para perguntas fechadas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateAlternativaDto"
            }
          }
        }
      },
...
    }
  }
}